Develop

+++ src/main/java/com/bndf1/test/calculator/domain/dto/OperandDTO.java
@@ -3,9 +3,11 @@ package com.bndf1.test.calculator.domain.dto;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @Builder(toBuilder = true)
 public class OperandDTO {
-  Double firstOperand;
-  Double secondOperand;
+  BigDecimal firstOperand;
+  BigDecimal secondOperand;
 }
